pr2/pyproject.toml
[project]
name = "pr2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdulbaseer102", email = "147842675+abdulbaseer102@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.4.302",
    "openai-agents>=0.0.7",
]

[project.scripts]
pr2 = "pr2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
